+++ CHANGELOG.md
@@ -3,6 +3,7 @@
 ## 16.07.1
 - added `Letter` as page size for PDF exports under Windows and OS X
   (for [Issue #234](https://github.com/pbek/QOwnNotes/issues/234))
+- added tooltips and a note count for items in the note sub folder tree
 - added more Dutch translation (thank you Telesight)
 
 ## 16.07.0